@@ -1,21 +1,39 @@
-group = 'tv.accedo.one'
-version = '1.1.2'
+plugins {
+    id 'com.android.library'
+    id 'maven-publish'
+}
 
-apply plugin: 'com.android.library'
+group = "tv.accedo.one"
+version = "1.2.1"
 
 android {
-    compileSdkVersion 29
+    namespace = "tv.accedo.one.sdk"
+    compileSdkVersion 33
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 29
+        minSdkVersion 21
+        targetSdkVersion 31
 
-        testInstrumentationRunnerArgument "accedo_one_test_apikey", project.hasProperty('accedo_one_test_apikey') ? "${accedo_one_test_apikey}" : ""
+        testInstrumentationRunnerArgument "accedo_one_test_apikey", project.hasProperty("accedo_one_test_apikey") ? "${accedo_one_test_apikey}" : ""
+    }
+
+    publishing {
+        singleVariant("release") {
+            withSourcesJar()
+        }
     }
 }
 
 dependencies {
+    implementation("androidx.annotation:annotation:1.5.0")
+    implementation("com.squareup.okhttp3:okhttp:4.10.0")
 
+    androidTestImplementation("junit:junit:4.13.2")
+    androidTestImplementation("androidx.test:core:1.5.0")
+    androidTestImplementation("androidx.test:runner:1.5.1")
+    androidTestImplementation("androidx.test:rules:1.5.0")
+    androidTestImplementation("androidx.test.ext:junit:1.1.4")
+    androidTestImplementation("androidx.test.ext:truth:1.5.0")
 }
 
-apply from: 'release-script.gradle'
+apply from: "release-script.gradle"
