@@ -6,7 +6,6 @@
 
 package tv.accedo.one.sdk;
 
-import android.support.test.runner.AndroidJUnit4;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -32,6 +31,9 @@ import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertNotNull;
 import static junit.framework.Assert.assertTrue;
 import static junit.framework.Assert.fail;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+
 
 @RunWith(AndroidJUnit4.class)
 public class AccedoOnePublishByAliasTest {
