@@ -6,7 +6,6 @@
 
 package tv.accedo.one.sdk;
 
-import android.support.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,6 +20,8 @@ import static tv.accedo.one.sdk.Shared.APP_ID;
 import static tv.accedo.one.sdk.Shared.DEVICE_ID;
 import static tv.accedo.one.sdk.Shared.getContext;
 import static junit.framework.Assert.assertTrue;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 @RunWith(AndroidJUnit4.class)
 public class AccedoOneResponseCheckerTest {
