@@ -7,7 +7,6 @@
 package tv.accedo.one.sdk;
 
 import android.content.Context;
-import android.support.test.runner.AndroidJUnit4;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -27,6 +26,8 @@ import tv.accedo.one.sdk.implementation.parsers.JSONMapByteParser;
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertTrue;
 import static junit.framework.Assert.fail;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 @RunWith(AndroidJUnit4.class)
 public class MockOneTest {
