@@ -2,6 +2,9 @@ package tv.accedo.one.sdk.definition;
 
 import android.content.Context;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import org.json.JSONObject;
 
 import java.util.Map;
@@ -12,45 +15,52 @@ import java.util.Map;
 public interface AccedoOneCache {
     /**
      * @param context
-     * @param key the metadata to fetch
+     * @param key     the metadata to fetch
      * @return the metadata requested by the key, or null if not yet fetched or empty.
      */
-    String getMetadata(Context context, String key);
+    @Nullable
+    String getMetadata(@NonNull Context context, @NonNull String key);
 
     /**
      * @param context
      * @return the list of all metadatas, or null if not yet fetched or empty.
      */
-    Map<String, String> getAllMetadata(Context context);
+    @Nullable
+    Map<String, String> getAllMetadata(@NonNull Context context);
 
     /**
      * @param context
      * @return the list of all metadatas as raw json, or null if not yet fetched or empty.
      */
-    JSONObject getAllMetadataRaw(Context context);
+    @Nullable
+    JSONObject getAllMetadataRaw(@NonNull Context context);
 
     /**
      * @param context
-     * @param key the asset to fetch
+     * @param key     the asset to fetch
      * @return the asset requested by the key, or null if not yet fetched or empty.
      */
-    byte[] getAsset(Context context, String key);
+    @Nullable
+    byte[] getAsset(@NonNull Context context, @NonNull String key);
 
     /**
      * @param context
      * @return the list of all asset URLS, or null if not yet fetched or empty.
      */
-    Map<String, String> getAllAssets(Context context);
+    @Nullable
+    Map<String, String> getAllAssets(@NonNull Context context);
 
     /**
      * @param context
      * @return The downloaded asset urls, and their content, or null if nothing has been fetched yet.
      */
-    Map<String, byte[]> getAllAssetsRaw(Context context);
+    @Nullable
+    Map<String, byte[]> getAllAssetsRaw(@NonNull Context context);
 
     /**
      * Removes all content from the cache.
+     *
      * @param context
      */
-    void clear(Context context);
+    void clear(@NonNull Context context);
 }
