@@ -1,5 +1,7 @@
 package tv.accedo.one.sdk.definition;
 
+import androidx.annotation.NonNull;
+
 import tv.accedo.one.sdk.definition.async.Cancellable;
 import tv.accedo.one.sdk.model.LogLevel;
 
@@ -12,7 +14,7 @@ public interface AccedoOneDetect {
      * @param message A text message for this log event. Messages should not exceed 10,000 characters. Longer messages will be truncated.
      * @param dimensions An optional list of up to four different dimensions. The possible values for each dimension must be predefined in the Accedo Onec Admin UI.
      */
-    void log(LogLevel logLevel, int code, String message, String... dimensions);
+    void log(@NonNull LogLevel logLevel, int code, @NonNull String message, String... dimensions);
 
     /**
      * Sends out all pending logs. Call this when leaving the app.
