@@ -13,10 +13,11 @@ class AccedoOneResponseChecker implements ResponseChecker<AccedoOneException>{
     public void throwIfNecessary(Response response) throws AccedoOneException {
         switch (response.getCode()){
             case -1: throw new AccedoOneException(StatusCode.NO_RESPONSE, response.getCaughtException());
-            case 200: break;
-            case 201: break;
-            case 204: break;
-            case 304: break;
+            case 200:
+            case 201:
+            case 204:
+            case 304:
+                break;
             case 400: throw new AccedoOneException(StatusCode.INVALID_PARAMETERS, response.getText(), response.getCaughtException());
             case 404: throw new AccedoOneException(StatusCode.KEY_NOT_FOUND, response.getText(), response.getCaughtException());
             default: throw new AccedoOneException(AccedoOneException.StatusCode.INVALID_RESPONSE, response.getText(), response.getCaughtException());
