@@ -3,6 +3,8 @@ package tv.accedo.one.sdk.implementation;
 import android.content.Context;
 import android.text.TextUtils;
 
+import androidx.annotation.NonNull;
+
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -12,10 +14,10 @@ import java.util.Map.Entry;
 import tv.accedo.one.sdk.definition.AccedoOneUserData;
 import tv.accedo.one.sdk.definition.async.AsyncAccedoOneUser;
 import tv.accedo.one.sdk.implementation.async.AsyncAccedoOneUserImpl;
-import tv.accedo.one.sdk.model.AccedoOneException;
-import tv.accedo.one.sdk.model.AccedoOneException.StatusCode;
 import tv.accedo.one.sdk.implementation.parsers.JSONMapParser;
 import tv.accedo.one.sdk.implementation.utils.Request.Method;
+import tv.accedo.one.sdk.model.AccedoOneException;
+import tv.accedo.one.sdk.model.AccedoOneException.StatusCode;
 
 /**
  * @author Pásztor Tibor Viktor <tibor.pasztor@accedo.tv>
@@ -30,23 +32,24 @@ public class AccedoOneUserDataImpl implements AccedoOneUserData {
         this.accedoOneImpl = accedoOneImpl;
     }
 
+    @NonNull
     @Override
-    public Map<String, String> getAllUserData(Context context, Scope scope, String userId) throws AccedoOneException {
+    public Map<String, String> getAllUserData(@NonNull Context context, @NonNull Scope scope, @NonNull String userId) throws AccedoOneException {
         //Input param check
-        if (TextUtils.isEmpty(userId) || scope == null) {
+        if (TextUtils.isEmpty(userId)) {
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS);
         }
 
         //Call
         return accedoOneImpl.createSessionedRestClient(accedoOneImpl.getEndpoint() + getPathForScope(scope) + userId)
-                .connect(new AccedoOneResponseChecker())
+                .connect(accedoOneImpl.okHttpClient, new AccedoOneResponseChecker())
                 .getParsedText(new JSONMapParser());
     }
 
     @Override
-    public void setAllUserData(Context context, Scope scope, String userId, Map<String, String> userData) throws AccedoOneException {
+    public void setAllUserData(@NonNull Context context, @NonNull Scope scope, @NonNull String userId, Map<String, String> userData) throws AccedoOneException {
         //Input param check
-        if (TextUtils.isEmpty(userId) || scope == null) {
+        if (TextUtils.isEmpty(userId)) {
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS);
         }
 
@@ -54,7 +57,7 @@ public class AccedoOneUserDataImpl implements AccedoOneUserData {
         JSONObject jsonObject = new JSONObject();
         try {
             if (userData != null) {
-                for(Entry<String, String> entry : userData.entrySet()) {
+                for (Entry<String, String> entry : userData.entrySet()) {
                     jsonObject.put(entry.getKey(), entry.getValue());
                 }
             }
@@ -64,44 +67,44 @@ public class AccedoOneUserDataImpl implements AccedoOneUserData {
 
         //Call
         accedoOneImpl.createSessionedRestClient(accedoOneImpl.getEndpoint() + getPathForScope(scope) + userId)
-                .setMethod(Method.POST)
+                .setMethod(Method.POST, jsonObject.toString())
                 .setHeader("Content-Type", "application/json; charset=utf-8")
-                .setPayload(jsonObject.toString())
-                .connect(new AccedoOneResponseChecker());
+                .connect(accedoOneImpl.okHttpClient, new AccedoOneResponseChecker());
     }
 
+    @NonNull
     @Override
-    public String getUserData(Context context, Scope scope, String userId, String key) throws AccedoOneException {
+    public String getUserData(@NonNull Context context, @NonNull Scope scope, @NonNull String userId, @NonNull String key) throws AccedoOneException {
         //Input param check
-        if (TextUtils.isEmpty(userId) || scope == null || key == null) {
+        if (userId.isEmpty()) {
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS);
         }
 
         //Call
         return accedoOneImpl.createSessionedRestClient(accedoOneImpl.getEndpoint() + getPathForScope(scope) + userId + "/" + key)
-                .connect(new AccedoOneResponseChecker())
+                .connect(accedoOneImpl.okHttpClient, new AccedoOneResponseChecker())
                 .getText();
     }
 
     @Override
-    public void setUserData(Context context, Scope scope, String userId, String key, String value) throws AccedoOneException {
+    public void setUserData(@NonNull Context context, @NonNull Scope scope, @NonNull String userId, @NonNull String key, @NonNull String value) throws AccedoOneException {
         //Input param check
-        if (TextUtils.isEmpty(userId) || key == null || value == null || scope == null) {
+        if (TextUtils.isEmpty(userId) || value.isEmpty()) {
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS);
         }
 
         //Call
         accedoOneImpl.createSessionedRestClient(accedoOneImpl.getEndpoint() + getPathForScope(scope) + userId + "/" + key)
-                .setMethod(Method.POST)
+                .setMethod(Method.POST, value)
                 .setHeader("Content-Type", "text/plain; charset=utf-8")
-                .setPayload(value)
-                .connect(new AccedoOneResponseChecker());
+                .connect(accedoOneImpl.okHttpClient, new AccedoOneResponseChecker());
     }
 
-    private String getPathForScope(Scope scope){
-        return Scope.APPLICATION.equals(scope)? PATH_USER_APPLICATION_DATA : PATH_USER_APPLICATIONGROUP_DATA;
+    private String getPathForScope(Scope scope) {
+        return Scope.APPLICATION.equals(scope) ? PATH_USER_APPLICATION_DATA : PATH_USER_APPLICATIONGROUP_DATA;
     }
 
+    @NonNull
     @Override
     public AsyncAccedoOneUser async() {
         return new AsyncAccedoOneUserImpl(this);
