@@ -2,6 +2,8 @@ package tv.accedo.one.sdk.implementation.mock;
 
 import android.content.Context;
 
+import androidx.annotation.NonNull;
+
 import org.json.JSONArray;
 import org.json.JSONObject;
 
@@ -31,8 +33,9 @@ public class MockOnePublish implements AccedoOnePublish {
         this.configSource = configSource;
     }
 
+    @NonNull
     @Override
-    public JSONObject getEntry(Context context, String id) throws AccedoOneException {
+    public JSONObject getEntry(@NonNull Context context, @NonNull String id) throws AccedoOneException {
         if (id == null) {
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS, "\"id\" can not be null.");
         }
@@ -40,13 +43,15 @@ public class MockOnePublish implements AccedoOnePublish {
         return PublishUtils.getEntryByMeta(getAllEntries(context), "id", id);
     }
 
+    @NonNull
     @Override
-    public JSONObject getEntry(Context context, String id, OptionalParams optionalParams) throws AccedoOneException {
+    public JSONObject getEntry(@NonNull Context context, @NonNull String id, OptionalParams optionalParams) throws AccedoOneException {
         return getEntry(context, id);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, String typeId) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull String typeId) throws AccedoOneException {
         if (typeId == null) {
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS, "\"typeId\" can not be null.");
         }
@@ -54,18 +59,21 @@ public class MockOnePublish implements AccedoOnePublish {
         return PublishUtils.getEntriesByMeta(getAllEntries(context), "typeId", typeId);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, String typeId, OptionalParams optionalParams) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull String typeId, OptionalParams optionalParams) throws AccedoOneException {
         return getEntries(context, typeId);
     }
 
+    @NonNull
     @Override
-    public PagedResponse getEntries(Context context, String typeId, PaginatedParams paginatedParams) throws AccedoOneException {
+    public PagedResponse getEntries(@NonNull Context context, @NonNull String typeId, PaginatedParams paginatedParams) throws AccedoOneException {
         return PublishUtils.filterByPage(getEntries(context, typeId), paginatedParams);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, List<String> ids) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull List<String> ids) throws AccedoOneException {
         if (ids == null || ids.isEmpty()) {
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS, "\"ids\" can not be null or empty.");
         }
@@ -73,33 +81,39 @@ public class MockOnePublish implements AccedoOnePublish {
         return PublishUtils.getEntriesByMetas(getAllEntries(context), "id", ids);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, List<String> ids, OptionalParams optionalParams) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull List<String> ids, OptionalParams optionalParams) throws AccedoOneException {
         return getEntries(context, ids);
     }
 
+    @NonNull
     @Override
-    public PagedResponse getEntries(Context context, List<String> ids, PaginatedParams paginatedParams) throws AccedoOneException {
+    public PagedResponse getEntries(@NonNull Context context, @NonNull List<String> ids, PaginatedParams paginatedParams) throws AccedoOneException {
         return PublishUtils.filterByPage(getEntries(context, ids), paginatedParams);
     }
 
+    @NonNull
     @Override
-    public JSONArray getAllEntries(Context context, OptionalParams optionalParams) throws AccedoOneException {
+    public JSONArray getAllEntries(@NonNull Context context, OptionalParams optionalParams) throws AccedoOneException {
         return getAllEntries(context);
     }
 
+    @NonNull
     @Override
-    public PagedResponse getAllEntries(Context context, PaginatedParams paginatedParams) throws AccedoOneException {
+    public PagedResponse getAllEntries(@NonNull Context context, PaginatedParams paginatedParams) throws AccedoOneException {
         return PublishUtils.filterByPage(getAllEntries(context), paginatedParams);
     }
 
+    @NonNull
     @Override
-    public List<PublishLocale> getAvailableLocales(Context context) throws AccedoOneException {
+    public List<PublishLocale> getAvailableLocales(@NonNull Context context) throws AccedoOneException {
         return new ArrayList<>();
     }
 
+    @NonNull
     @Override
-    public JSONArray getAllEntries(Context context) throws AccedoOneException {
+    public JSONArray getAllEntries(@NonNull Context context) throws AccedoOneException {
         if (entries == null) {
             try {
                 entries = configSource.getAllEntries(context);
@@ -111,11 +125,13 @@ public class MockOnePublish implements AccedoOnePublish {
         return entries;
     }
 
+    @NonNull
     @Override
     public AsyncAccedoOnePublish async() {
         return new AsyncAccedoOnePublishImpl(this);
     }
 
+    @NonNull
     @Override
     public AccedoOnePublishByAlias byAlias() {
         return new MockOnePublishByAlias(this);
