@@ -2,6 +2,8 @@ package tv.accedo.one.sdk.implementation.mock;
 
 import android.content.Context;
 
+import androidx.annotation.NonNull;
+
 import org.json.JSONArray;
 import org.json.JSONObject;
 
@@ -26,8 +28,9 @@ public class MockOnePublishByAlias implements AccedoOnePublishByAlias {
         this.mockOnePublish = mockOnePublish;
     }
 
+    @NonNull
     @Override
-    public JSONObject getEntry(Context context, String alias) throws AccedoOneException {
+    public JSONObject getEntry(@NonNull Context context, @NonNull String alias) throws AccedoOneException {
         if(alias==null){
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS, "\"alias\" can not be null.");
         }
@@ -35,13 +38,15 @@ public class MockOnePublishByAlias implements AccedoOnePublishByAlias {
         return PublishUtils.getEntryByMeta(mockOnePublish.getAllEntries(context), "entryAlias", alias);
     }
 
+    @NonNull
     @Override
-    public JSONObject getEntry(Context context, String alias, OptionalParams optionalParams) throws AccedoOneException {
+    public JSONObject getEntry(@NonNull Context context, @NonNull String alias, OptionalParams optionalParams) throws AccedoOneException {
         return getEntry(context, alias);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, String typeAlias) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull String typeAlias) throws AccedoOneException {
         if(typeAlias==null){
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS, "\"typeAlias\" can not be null.");
         }
@@ -49,18 +54,21 @@ public class MockOnePublishByAlias implements AccedoOnePublishByAlias {
         return PublishUtils.getEntriesByMeta(mockOnePublish.getAllEntries(context), "typeAlias", typeAlias);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, String typeAlias, OptionalParams optionalParams) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull String typeAlias, OptionalParams optionalParams) throws AccedoOneException {
         return getEntries(context, typeAlias);
     }
 
+    @NonNull
     @Override
-    public PagedResponse getEntries(Context context, String typeAlias, PaginatedParams paginatedParams) throws AccedoOneException {
+    public PagedResponse getEntries(@NonNull Context context, @NonNull String typeAlias, PaginatedParams paginatedParams) throws AccedoOneException {
         return PublishUtils.filterByPage(getEntries(context, typeAlias), paginatedParams);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, List<String> aliases) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull List<String> aliases) throws AccedoOneException {
         if(aliases==null || aliases.isEmpty()){
             throw new AccedoOneException(StatusCode.INVALID_PARAMETERS, "\"aliases\" can not be null or empty.");
         }
@@ -68,16 +76,19 @@ public class MockOnePublishByAlias implements AccedoOnePublishByAlias {
         return PublishUtils.getEntriesByMetas(mockOnePublish.getAllEntries(context), "entryAlias", aliases);
     }
 
+    @NonNull
     @Override
-    public JSONArray getEntries(Context context, List<String> aliases, OptionalParams optionalParams) throws AccedoOneException {
+    public JSONArray getEntries(@NonNull Context context, @NonNull List<String> aliases, OptionalParams optionalParams) throws AccedoOneException {
         return getEntries(context, aliases);
     }
 
+    @NonNull
     @Override
-    public PagedResponse getEntries(Context context, List<String> aliases, PaginatedParams paginatedParams) throws AccedoOneException {
+    public PagedResponse getEntries(@NonNull Context context, @NonNull List<String> aliases, PaginatedParams paginatedParams) throws AccedoOneException {
         return PublishUtils.filterByPage(getEntries(context, aliases), paginatedParams);
     }
 
+    @NonNull
     @Override
     public AsyncAccedoOnePublishByAlias async() {
         return new AsyncAccedoOnePublishByAliasImpl(this);
