@@ -2,6 +2,8 @@ package tv.accedo.one.sdk.implementation.parsers;
 
 import org.json.JSONObject;
 
+import java.nio.charset.StandardCharsets;
+
 import tv.accedo.one.sdk.implementation.utils.Response.ThrowingParser;
 import tv.accedo.one.sdk.model.AccedoOneException;
 import tv.accedo.one.sdk.model.AccedoOneException.StatusCode;
@@ -13,7 +15,7 @@ public class JSONObjectByteParser implements ThrowingParser<byte[], JSONObject, 
     @Override
     public JSONObject parse(byte[] input) throws AccedoOneException {
         try {
-            return new JSONObject(new String(input, "UTF-8"));
+            return new JSONObject(new String(input, StandardCharsets.UTF_8));
         } catch (Exception e) {
             throw new AccedoOneException(StatusCode.INVALID_RESPONSE, e);
         }
