@@ -8,21 +8,23 @@ package tv.accedo.one.sdk.implementation.utils;
 
 import android.util.Log;
 
-import java.io.OutputStream;
-import java.io.UnsupportedEncodingException;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import java.net.HttpCookie;
-import java.net.HttpURLConnection;
-import java.net.URL;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
+
+import okhttp3.OkHttpClient;
+import okhttp3.RequestBody;
 
 /**
  * @author Pásztor Tibor Viktor <tibor.pasztor@accedo.tv>
  */
 public class Request {
     public static final String TAG = "RestClient";
-    protected static final int DEFAULT_TIMEOUT_CONNECT = 5000;
-    protected static final int DEFAULT_TIMEOUT_READ = 10000;
 
     public static enum Method {
         GET,
@@ -32,12 +34,14 @@ public class Request {
     }
 
     protected String url;
-    protected HttpURLConnection httpUrlConnection;
-    protected Exception caughtCreationException;
-    protected byte[] payload;
-    protected String charset = "UTF-8";
-    protected ArrayList<HttpCookie> cookies = new ArrayList<HttpCookie>();
 
+    okhttp3.Request.Builder requestBuilder = new okhttp3.Request.Builder();
+
+    protected Exception caughtCreationException;
+    @NonNull
+    protected Charset charset = StandardCharsets.UTF_8;
+    protected ArrayList<HttpCookie> cookies = new ArrayList<HttpCookie>();
+    @Nullable
     protected OnResponseListener onResponseListener;
 
     /**
@@ -53,9 +57,13 @@ public class Request {
      * @param method GET, POST, PUT or DELETE
      * @return this RestClient instance for chaining
      */
-    public Request setMethod(Method method) {
+    public Request setMethod(Method method, @Nullable String payload) {
         try {
-            httpUrlConnection.setRequestMethod(method.name());
+            RequestBody body = null;
+            if (payload != null && payload.isEmpty()) {
+                body = RequestBody.create(payload.getBytes(charset));
+            }
+            requestBuilder.method(method.name(), body);
         } catch (Exception e) {
             Utils.log(e);
         }
@@ -70,8 +78,8 @@ public class Request {
      * @return this RestClient instance for chaining
      */
     public Request setHeader(String key, String value) {
-        if (httpUrlConnection != null) {
-            httpUrlConnection.setRequestProperty(key, value);
+        if (requestBuilder != null) {
+            requestBuilder.header(key, value);
         }
         return this;
     }
@@ -84,8 +92,8 @@ public class Request {
      * @return this RestClient instance for chaining
      */
     public Request addHeader(String key, String value) {
-        if (httpUrlConnection != null) {
-            httpUrlConnection.addRequestProperty(key, value);
+        if (requestBuilder != null) {
+            requestBuilder.addHeader(key, value);
         }
         return this;
     }
@@ -123,57 +131,10 @@ public class Request {
         }
 
         //Set to urlConnection
-        if (httpUrlConnection != null) {
-            httpUrlConnection.setRequestProperty("Cookie", cookieString.toString());
+        if (requestBuilder != null) {
+            requestBuilder.addHeader("Cookie", cookieString.toString());
         }
 
-        return this;
-    }
-
-    /**
-     * Sets the payload to be sent in the request body. The string will be processed with the charset active at the time this value is set.
-     * !!To make sure that this value is encoded properly, make sure the charset is set before calling this method.!!
-     *
-     * @param output The string to be sent in the request body.
-     * @return this RestClient instance for chaining
-     */
-    public Request setPayload(String payload) {
-        if (payload != null) {
-            try {
-                this.payload = payload.getBytes(charset);
-            } catch (UnsupportedEncodingException e) {
-                Utils.log(e);
-            }
-        } else {
-            this.payload = null;
-        }
-
-        return this;
-    }
-
-    /**
-     * Sets the payload to be sent in the request body.
-     *
-     * @param rawOutput The data to be sent in the request body.
-     * @return this RestClient instance for chaining
-     */
-    public Request setPayload(byte[] payload) {
-        this.payload = payload;
-        return null;
-    }
-
-    /**
-     * Sets the timeouts for the request.
-     *
-     * @param connect the timeout value for the connection to be established
-     * @param read    the timeout value for the fetching operation to be completed
-     * @return this RestClient instance for chaining
-     */
-    public Request setTimeout(int connect, int read) {
-        if (httpUrlConnection != null) {
-            httpUrlConnection.setConnectTimeout(connect);
-            httpUrlConnection.setReadTimeout(read);
-        }
         return this;
     }
 
@@ -183,7 +144,7 @@ public class Request {
      * @param charset the RestClient instance should use. Doesn't accept null.
      * @return this RestClient instance for chaining
      */
-    public Request setCharset(String charset) {
+    public Request setCharset(Charset charset) {
         if (charset != null) {
             this.charset = charset;
         }
@@ -204,30 +165,24 @@ public class Request {
     /**
      * Creates a non-cached RestClient instance.
      *
-     * @param url          the url to load
-     * @param httpCacheDir the folder to use for httpCaching. Recommended: new File(context.getCacheDir(), "http")
+     * @param url the url to load
      */
-    public Request(String url) {
+    public Request(@NonNull String url) {
         this.url = url;
 
         //Make HttpUrlConnection
         try {
-            httpUrlConnection = (HttpURLConnection) (new URL(url).openConnection());
+            requestBuilder = new okhttp3.Request.Builder().url(url);
         } catch (Exception e) {
             caughtCreationException = e;
             Utils.log(e);
         }
 
-        //Init httpUrlConnection
-        if (httpUrlConnection != null) {
-            httpUrlConnection.setConnectTimeout(DEFAULT_TIMEOUT_CONNECT);
-            httpUrlConnection.setReadTimeout(DEFAULT_TIMEOUT_READ);
-            httpUrlConnection.setUseCaches(false);
-        }
+
     }
 
-    public <E extends Exception> Response connect(ResponseChecker<E> responseChecker) throws E {
-        Response response = fetchResponse();
+    public <E extends Exception> Response connect(@NonNull OkHttpClient okHttpClient, ResponseChecker<E> responseChecker) throws E {
+        Response response = fetchResponse(okHttpClient);
 
         //Check response, and throw if necessary
         if (responseChecker != null) {
@@ -242,31 +197,27 @@ public class Request {
         return response;
     }
 
-    private Response fetchResponse() {
+    private Response fetchResponse(@NonNull OkHttpClient okHttpClient) {
         Response response = new Response(url, caughtCreationException);
-        OutputStream os = null;
 
-        if (caughtCreationException == null && httpUrlConnection != null && httpUrlConnection.getURL() != null) {
+        okhttp3.Request req = requestBuilder.build();
+
+        if (caughtCreationException == null) {
+            req.url();
             try {
                 //Logging
-                Utils.log(Log.DEBUG, "Sending " + httpUrlConnection.getRequestMethod() + " request: " + httpUrlConnection.getURL().toString());
+                Utils.log(Log.DEBUG, "Sending " + req.method() + " request: " + req.url());
 
                 //Request
-                httpUrlConnection.connect();
-                if (payload != null) {
-                    os = httpUrlConnection.getOutputStream();
-                    os.write(payload);
-                    os.flush();
-                    os.close();
-                }
-                response = new Response(httpUrlConnection, url, charset);
+                okhttp3.Response okHttpResponse = okHttpClient.newCall(req).execute();
+
+                response = new Response(okHttpResponse, url, charset);
 
             } catch (Exception e) {
                 Utils.log(e);
                 response = new Response(url, e);
-            } finally {
-                Utils.closeQuietly(os);
             }
+
         }
 
         return response;
